update: code stabilty

+++ streamlit_app.py
@@ -1,4 +1,5 @@
 import base64
+import json
 import os
 import requests
 import streamlit as st
@@ -41,39 +42,43 @@ elif sel_menu == "View criminals":
     submit_url = f"{SERVER_URL}/api/list"
     
     st.header("View criminals")
-    try:
-        st.warning("Loading list of criminals...")
+    # try:
+    st.warning("Loading list of criminals...")
+    
+    res = requests.get(submit_url, data={"authKey": AUTH_KEY}, timeout=10)
+    res_json = res.json()
+    
+    if res_json.get('error', True):
+        st.error(f"Loading failed: {res_json.get('detail', 'Unknown error')}")
+    else:
+        st.success(f"Loading successful: {res_json.get('detail', 'Unknown result')}")
         
-        res = requests.get(submit_url, data={"authKey": AUTH_KEY}, timeout=10)
-        res_json = res.json()
+        res_json = res_json['data']
+        if res_json == []:
+            st.write("No criminals registered.")
         
-        if res_json.get('error', True):
-            st.error(f"Loading failed: {res_json.get('detail', 'Unknown error')}")
         else:
-            st.success(f"Loading successful: {res_json.get('detail', 'Unknown result')}")
-            
-            res_json = res_json['data']
-            if res_json == []:
-                st.write("No criminals registered.")
-                
-            else:
-                crimi_layout_cell = []
-                for i in range((len(res_json) + 2) // 3):
-                    crimi_layout_cell.append(st.columns(3))
+            crimi_layout_cell = []
+            for i in range((len(res_json) + 2) // 3):
+                crimi_layout_cell.append(st.columns(3))
 
-                crimi_layout_cell = [col for row in crimi_layout_cell for col in row]
+            crimi_layout_cell = [col for row in crimi_layout_cell for col in row]
 
-                for index, criminal in enumerate(res_json):
-                    crimi_layout_cell[index].write(f"Name: {criminal['crimi_name']}")
-                    crimi_layout_cell[index].write(f"Desc: {criminal['crimi_desc']}")
+            for index, criminal in enumerate(res_json):
+                crimi_layout_cell[index].write(f"Name: {criminal['crimi_name']}")
+                crimi_layout_cell[index].write(f"Desc: {criminal['crimi_desc']}")
+                if criminal.get('crimi_face', None) != None:
                     with open(os.path.join(TEMP_DIR, 'view_temp.jpg'), 'wb') as f:
                         f.write(base64.b64decode(criminal['crimi_face'][0]))
                     crimi_layout_cell[index].image(os.path.join(TEMP_DIR, 'view_temp.jpg'), caption=f"Regi time: {criminal['regi_time']}", use_column_width=True)
                     os.remove(os.path.join(TEMP_DIR, 'view_temp.jpg'))
-                    crimi_layout_cell[index].write("")
+                else:
+                    crimi_layout_cell[index].write("No image")
+                
+                crimi_layout_cell[index].write("")
         
-    except Exception as e:
-        st.error("Failed to load criminals. Are you connected to the Internet?")
+    # except Exception as e:
+        # st.error("Failed to load criminals. Are you connected to the Internet?")
 
 # Search Criminal
 elif sel_menu == "Search Criminal":
@@ -101,47 +106,53 @@ elif sel_menu == "Search Criminal":
         else:
             st.warning("Searching...")
             
-            try:
-                res = requests.post(submit_url, data={'authKey': AUTH_KEY}, files={"image": criminal_image}, timeout=10)
-                res_json = res.json()
+            # try:
+            res = requests.post(submit_url, data={'authKey': AUTH_KEY}, files={"image": criminal_image}, timeout=10)
+            res_json = res.json()
+            
+            if res_json.get("error", True):
+                st.error(f"Search failed: {res_json.get("detail", "Unknown error")}")
+            else:
+                st.success(f"Search successful: {res_json.get("detail", "Unknown result")}")
                 
-                if res_json.get("error", True):
-                    st.error(f"Search failed: {res_json.get("detail", "Unknown error")}")
-                else:
-                    st.success(f"Search successful: {res_json.get("detail", "Unknown result")}")
-                    
-                    similar_layout_cell = []
-                    for i in range((len(res_json["similar_info"]) + 2 + 3) // 3):
-                        similar_layout_cell.append(st.columns(3))
-                    similar_layout_cell = [col for row in similar_layout_cell for col in row]
-                    
-                    suspect_layout_cell = []
-                    for i in range((len(res_json["suspect_info"]) + 2 + 3) // 3):
-                        suspect_layout_cell.append(st.columns(3))
-                    suspect_layout_cell = [col for row in suspect_layout_cell for col in row]
-                    
-                    similar_layout_cell[0].header("Similar")
-                    for index, criminal in enumerate(res_json["similar_info"]):
-                        similar_layout_cell[index+3].write(f"Name: {criminal['crimi_name']}")
-                        similar_layout_cell[index+3].write(f"Desc: {criminal['crimi_desc']}")
+                similar_layout_cell = []
+                for i in range((len(res_json["similar_info"]) + 2 + 3) // 3):
+                    similar_layout_cell.append(st.columns(3))
+                similar_layout_cell = [col for row in similar_layout_cell for col in row]
+                
+                suspect_layout_cell = []
+                for i in range((len(res_json["suspect_info"]) + 2 + 3) // 3):
+                    suspect_layout_cell.append(st.columns(3))
+                suspect_layout_cell = [col for row in suspect_layout_cell for col in row]
+                
+                similar_layout_cell[0].header("Similar")
+                for index, simil_crimi in enumerate(res_json["similar_info"]):
+                    similar_layout_cell[index+3].write(f"Name: {simil_crimi['crimi_name']}")
+                    similar_layout_cell[index+3].write(f"Desc: {simil_crimi['crimi_desc']}")
+                    if simil_crimi.get('crimi_face', None) != None:
                         with open(os.path.join(TEMP_DIR, 'search_temp.jpg'), 'wb') as f:
-                            f.write(base64.b64decode(criminal['crimi_face']))
+                            f.write(base64.b64decode(simil_crimi['crimi_face']))
                         similar_layout_cell[index+3].image(os.path.join(TEMP_DIR, 'search_temp.jpg'), use_column_width=True)
                         os.remove(os.path.join(TEMP_DIR, 'search_temp.jpg'))
-                        similar_layout_cell[index+3].write("")
-                    
-                    suspect_layout_cell[0].header("Suspect")
-                    for index, criminal in enumerate(res_json["suspect_info"]):
-                        suspect_layout_cell[index+3].write(f"Name: {criminal['crimi_name']}")
-                        suspect_layout_cell[index+3].write(f"Desc: {criminal['crimi_desc']}")
+                    else:
+                        similar_layout_cell[index+3].write("No image")
+                    similar_layout_cell[index+3].write("")
+                
+                suspect_layout_cell[0].header("Suspect")
+                for index, sus_crimi in enumerate(res_json["suspect_info"]):
+                    suspect_layout_cell[index+3].write(f"Name: {sus_crimi['crimi_name']}")
+                    suspect_layout_cell[index+3].write(f"Desc: {sus_crimi['crimi_desc']}")
+                    if sus_crimi.get('crimi_face', None) != None:
                         with open(os.path.join(TEMP_DIR, 'search_temp.jpg'), 'wb') as f:
-                            f.write(base64.b64decode(criminal['crimi_face']))
+                            f.write(base64.b64decode(sus_crimi['crimi_face']))
                         suspect_layout_cell[index+3].image(os.path.join(TEMP_DIR, 'search_temp.jpg'), use_column_width=True)
                         os.remove(os.path.join(TEMP_DIR, 'search_temp.jpg'))
-                        suspect_layout_cell[index+3].write("")
+                    else:
+                        suspect_layout_cell[index+3].write("No image")
+                    suspect_layout_cell[index+3].write("")
                 
-            except Exception as e:
-                st.error(f"Search failed. Are you connected to the Internet?: {e}")
+            # except Exception as e:
+            #     st.error(f"Search failed. Are you connected to the Internet?: {e}")
             
 # Regi Criminal
 elif sel_menu == "Regi Criminal":
